Moved AggChildHealthAggregationHelper to own file
